+++ src/main/java/com/jcraft/jsch/jce/SignatureECDSAN.java
@@ -142,8 +142,8 @@ public abstract class SignatureECDSAN implements com.jcraft.jsch.SignatureECDSA 
       byte[] r = b.getMPInt();
       byte[] s = b.getMPInt();
 
-      r=insert0(r);
-      s=insert0(s);
+      r=trimLeadingZeros(insert0(r));
+      s=trimLeadingZeros(insert0(s));
 
       byte[] asn1 = null;
       if(r.length<64){
@@ -175,14 +175,15 @@ public abstract class SignatureECDSAN implements com.jcraft.jsch.SignatureECDSA 
     return signature.verify(sig); 
   }
 
-  private byte[] insert0(byte[] buf){
+  private static byte[] insert0(byte[] buf){
     if ((buf[0] & 0x80) == 0) return buf;
     byte[] tmp = new byte[buf.length+1];
     System.arraycopy(buf, 0, tmp, 1, buf.length);
     bzero(buf);
     return tmp;
   }
-  private byte[] chop0(byte[] buf){
+
+  private static byte[] chop0(byte[] buf){
     if(buf[0]!=0) return buf;
     byte[] tmp = new byte[buf.length-1];
     System.arraycopy(buf, 1, tmp, 0, tmp.length);
@@ -190,7 +191,24 @@ public abstract class SignatureECDSAN implements com.jcraft.jsch.SignatureECDSA 
     return tmp;
   }
 
-  private void bzero(byte[] buf){
+  private static void bzero(byte[] buf){
     for(int i = 0; i<buf.length; i++) buf[i]=0;
+  }
+
+  private static byte[] trimLeadingZeros(byte[] buf){
+    if(buf.length<2) return buf;
+
+    int i=0;
+    while(i<buf.length-1){
+      if(buf[i] == 0 && (buf[i+1] & 0x80) == 0) i++;
+      else break;
+    }
+
+    if(i == 0) return buf;
+
+    byte[] tmp = new byte[buf.length-i];
+    System.arraycopy(buf, i, tmp, 0, tmp.length);
+    bzero(buf);
+    return tmp;
   }
 }